backend/requirements.txt
fastapi
uvicorn[standard]
requests
numpy
faiss-cpu
pydantic
